refactor(summary): extract common components
