+++ projects/client/src/lib/sections/summary/components/media/v2/_internal/MediaActions.svelte
@@ -6,6 +6,7 @@
   import ListDropdown from "$lib/sections/summary/components/list-dropdown/ListDropdown.svelte";
   import type { ListDropdownProps } from "$lib/sections/summary/components/list-dropdown/ListDropdownProps";
   import { useWatchCount } from "$lib/stores/useWatchCount";
+  import SummaryActions from "../../../_internal/SummaryActions.svelte";
   import MediaActionsPopupMenu from "./MediaActionsPopupMenu.svelte";
   import { TrackIntlProvider } from "./TrackIntlProvider";
 
@@ -33,40 +34,8 @@
   });
 </script>
 
-<div class="trakt-media-actions">
+<SummaryActions>
   <MarkAsWatchedAction {...markAsWatchedProps} i18n={TrackIntlProvider} />
   <ListDropdown {...listProps} style="popup" />
   <MediaActionsPopupMenu {media} {streamOn} {title} />
-</div>
-
-<style>
-  .trakt-media-actions {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    gap: var(--gap-xs);
-
-    position: relative;
-
-    height: var(--ni-56);
-    width: var(--ni-272);
-
-    padding: var(--ni-8) var(--ni-10);
-    box-sizing: border-box;
-
-    background-color: var(--color-card-background);
-    border-radius: var(--border-radius-l);
-    box-shadow: var(--popup-shadow);
-
-    :global(.trakt-button) {
-      flex-grow: 1;
-
-      flex-direction: row-reverse;
-      justify-content: flex-end;
-    }
-
-    :global(.trakt-popup-menu-button) {
-      color: var(--color-text-primary);
-    }
-  }
-</style>
+</SummaryActions>
